+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/icon/DefaultFontImageIcon.java
@@ -70,8 +70,8 @@ final class DefaultFontImageIcon implements FontImageIcon {
 
 		private final Ikon ikon;
 
-		private int size = Icons.ICON_SIZE.getOrThrow();
-		private Color color = Icons.ICON_COLOR.getOrThrow();
+		private int size = Icons.SIZE.getOrThrow();
+		private Color color = Icons.COLOR.getOrThrow();
 		private IconPainter iconPainter = DEFAULT_ICON_PAINTER;
 		private ImageIconFactory imageIconFactory = DEFAULT_ICON_FACTORY;