+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/icon/package-info.java
@@ -19,8 +19,8 @@
 /**
  * Package configuration values:
  * <ul>
- * <li>{@link is.codion.swing.common.ui.icon.Icons#ICON_SIZE}
- * <li>{@link is.codion.swing.common.ui.icon.Icons#ICON_COLOR}
+ * <li>{@link is.codion.swing.common.ui.icon.Icons#SIZE}
+ * <li>{@link is.codion.swing.common.ui.icon.Icons#COLOR}
  * </ul>
  */
 @org.jspecify.annotations.NullMarked